+++ burner-wallet-plugin/tokenbridge-plugin/src/burner-wallet/assets/Etc.ts
