+++ burner-wallet-plugin/tokenbridge-plugin/src/burner-wallet/index.ts
@@ -1,0 +1,6 @@
+export { default as sPOA } from './assets/sPOA'
+export { default as Etc } from './assets/Etc'
+export { default as Wetc } from './assets/Wetc'
+export { default as ERC677Asset } from './assets/ERC677Asset'
+export { default as TokenBridgeGateway } from './gateways/TokenBridgeGateway'
+export { default as Bridge } from './pairs/Bridge'
