+++ burner-wallet-plugin/tokenbridge-plugin/src/index.ts
@@ -1,7 +1,2 @@
-export { default as sPOA } from './assets/sPOA'
-export { default as Etc } from './assets/Etc'
-export { default as Wetc } from './assets/Wetc'
-export { default as ERC677Asset } from './assets/ERC677Asset'
-export { default as TokenBridgeGateway } from './gateways/TokenBridgeGateway'
-export { default as WETCBridge } from './pairs/WETCBridge'
-export { default as Bridge } from './pairs/Bridge'
+export { ERC677Asset, sPOA, Etc, Wetc, TokenBridgeGateway, Bridge } from './burner-wallet'
+export { WETCBridge } from './wetc-bridge'
