+++ burner-wallet-plugin/tokenbridge-plugin/src/wetc-bridge/index.ts
@@ -1,0 +1,1 @@
+export { default as WETCBridge } from './WETCBridge'
